Fixed Instruction class for long conversion

+++ src/com/kennesaw/cpumodule/CPU.java
@@ -21,7 +21,7 @@ public class CPU {
     public void runProcess() {
         while(isRunning) {
             // Get instruction from memory
-            int instr = fetch(cpuState.getPc());
+            long instr = fetch(cpuState.getPc());
 
             // Increment the PC
             cpuState.incrementPc();
@@ -51,12 +51,12 @@ public class CPU {
         }
     }
 
-    private int fetch(int addr) {
+    private long fetch(int addr) {
         addr *= 4;
         return dmaChannel.readRAM(addr);
     }
 
-    private void decode(int instructionBin) {
+    private void decode(long instructionBin) {
         cpuState.setInstruction(instructionBin);
     }
 
@@ -67,37 +67,37 @@ public class CPU {
                 cpuState.setReg(s1, cpuState.getReg(s2));
                 break;
             case 0x05:
-                acc = ALU.add(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.add((int) cpuState.getReg(s1), (int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x06:
-                acc = ALU.sub(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.sub((int) cpuState.getReg(s1),(int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x07:
-                acc = ALU.mult(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.mult((int) cpuState.getReg(s1),(int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x08:
-                acc = ALU.div(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.div((int) cpuState.getReg(s1),(int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x09:
-                acc = ALU.and(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.and((int) cpuState.getReg(s1), (int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x0A:
-                acc = ALU.or(cpuState.getReg(s1),cpuState.getReg(s2));
+                acc = ALU.or((int) cpuState.getReg(s1), (int) cpuState.getReg(s2));
                 //cpuState.setReg(State.ACCUMULATOR_REG, acc);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x10:
-                if (ALU.lessThan(cpuState.getReg(s1),cpuState.getReg(s2))) cpuState.setReg(dr, 1);
+                if (ALU.lessThan((int) cpuState.getReg(s1), (int) cpuState.getReg(s2))) cpuState.setReg(dr, 1);
                 else cpuState.setReg(dr, 0);
                 break;
         }
@@ -107,41 +107,41 @@ public class CPU {
         int acc;
         switch (opcode) {
             case 0x02:
-                dmaChannel.writeRAM(cpuState.getReg(dr), cpuState.getReg(br));
+                dmaChannel.writeRAM((int) cpuState.getReg(dr), cpuState.getReg(br));
                 break;
             case 0x03:
-                cpuState.setReg(dr, dmaChannel.readRAM(cpuState.getReg(br)));
+                cpuState.setReg(dr, dmaChannel.readRAM((int) cpuState.getReg(br)));
                 break;
             case 0x0B:
                 cpuState.setReg(dr, addr);
                 break;
             case 0x0C:
-                acc = ALU.add(cpuState.getReg(dr), addr);
+                acc = ALU.add((int) cpuState.getReg(dr), addr);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x0D:
-                acc = ALU.mult(cpuState.getReg(dr), addr);
+                acc = ALU.mult((int) cpuState.getReg(dr), addr);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x0E:
-                acc = ALU.div(cpuState.getReg(dr), addr);
+                acc = ALU.div((int) cpuState.getReg(dr), addr);
                 cpuState.setReg(dr, acc);
                 break;
             case 0x0F:
                 cpuState.setReg(dr, addr);
                 break;
             case 0x11:
-                if (ALU.lessThan(cpuState.getReg(br), addr)) cpuState.setReg(dr, 1);
+                if (ALU.lessThan((int) cpuState.getReg(br), addr)) cpuState.setReg(dr, 1);
                 else cpuState.setReg(dr, 0);
                 break;
             case 0x15:
-                if (ALU.isBranchEqualTo(cpuState.getReg(br), cpuState.getReg(dr))) cpuState.setPc(addr/4);
+                if (ALU.isBranchEqualTo((int) cpuState.getReg(br), (int) cpuState.getReg(dr))) cpuState.setPc(addr/4);
                 break;
             case 0x16:
-                if (ALU.isBranchInnequalTo(cpuState.getReg(br), cpuState.getReg(dr))) cpuState.setPc(addr/4);
+                if (ALU.isBranchInnequalTo((int) cpuState.getReg(br), (int) cpuState.getReg(dr))) cpuState.setPc(addr/4);
                 break;
             case 0x17:
-                if (ALU.isBranchNotZero(cpuState.getReg(br))) cpuState.setPc(addr/4);
+                if (ALU.isBranchNotZero((int) cpuState.getReg(br))) cpuState.setPc(addr/4);
                 break;
             case 0x18:
                 break;
@@ -166,12 +166,12 @@ public class CPU {
     private void executeIO(byte opcode, byte r1, byte r2, int addr) {
         switch (opcode) {
             case 0x00:
-                if (addr == 0) cpuState.setReg(r1, dmaChannel.readRAM(cpuState.getReg(r2)));
+                if (addr == 0) cpuState.setReg(r1, dmaChannel.readRAM((int) cpuState.getReg(r2)));
                 else cpuState.setReg(r1, dmaChannel.readRAM(addr));
                 break;
             case 0x01:
                 if (r1 == r2) dmaChannel.writeRAM(addr, cpuState.getReg(r1));
-                else dmaChannel.writeRAM(cpuState.getReg(r2), cpuState.getReg(r1));
+                else dmaChannel.writeRAM((int) cpuState.getReg(r2), cpuState.getReg(r1));
                 break;
         }
     }
